+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog - [0.2.0] - yyyy-mm-dd
 
 ### New Features
-
+1. Project all (non-string) node properties using the appropriate (integer/float) types.
 
 
 ### Bug Fixies